+++ sympy_extras/integrals/__init__.py
@@ -17,6 +17,7 @@ from .dfinite import chyzak, dfinite_ode, dfinite_integral, DFiniteTelescoper
 from .asymptotic import asymptotic_integral, watson_lemma, laplace_method, stationary_phase
 from .elliptic import elliptic_integral
 from .series import series_integral
+from .tables import table_integral, lookup, TABLE, TableEntry
 from .risch import risch_antiderivative, is_nonelementary
 from .parametric import parametric_integral
 from .brackets import ramanujan_master_theorem, method_of_brackets, mellin_transform_series
@@ -31,4 +32,4 @@ __all__ = ['ConditionalValue', 'GammaQuotient', 'MellinTransform', 'mellin_trans
     'is_nonelementary', 'principal_value_integral', 'mean_value_integral', 'laurent_coefficient', 'contour_integral', 'rectangular_integral',
     'sector_integral', 'indented_integral', 'frullani', 'glasser', 'transformation_integral', 'laplace_rules', 'laplace_integral',
     'laplace_of_convolution', 'chyzak', 'dfinite_ode', 'dfinite_integral', 'DFiniteTelescoper', 'asymptotic_integral', 'watson_lemma', 'laplace_method',
-    'stationary_phase', 'elliptic_integral', 'series_integral']
+    'stationary_phase', 'elliptic_integral', 'series_integral', 'table_integral', 'lookup', 'TABLE', 'TableEntry']